TrabalhoServerEntities/src/main/java/br/edu/ifes/sr/poo2/model/Categoria.java
package br.edu.ifes.sr.poo2.model;

import javax.persistence.Column;
import javax.persistence.Entity;

@Entity
public class Categoria extends Model {

	/**
	 * 
	 */
	private static final long serialVersionUID = 1L;

	@Column
	private String nome;
	
	@Column
	private String descricao;

	public String getNome() {
		return nome;
	}

	public void setNome(String nome) {
		this.nome = nome;
	}

	public String getDescricao() {
		return descricao;
	}

	public void setDescricao(String descricao) {
		this.descricao = descricao;
	}
	
}
